requirements.txt
pyjsonselect
nose
pypandoc
